mem0/examples/mem0_persistent_graphMemory/chatbot/requirements.txt
openai
python-dotenv
mem0ai[graph]
git+https://github.com/langchain-ai/langchain-memgraph.git@main
